clear 13

+++ model/todo.go
@@ -5,7 +5,7 @@ import "time"
 type (
 	// A TODO expresses ...
 	TODO struct {
-		ID          int       `json:"id"`
+		ID          int64     `json:"id"`
 		Subject     string    `json:"subject"`
 		Description string    `json:"description"`
 		CreatedAt   time.Time `json:"created_at"`
@@ -27,9 +27,9 @@ type (
 	// A ReadTODOResponse expresses ...
 	ReadTODOResponse struct{}
 
-	// A UpdateTODORequest expresses ...
+	// A UpdateTODORequest expresses ...｀｀
 	UpdateTODORequest struct {
-		ID          int    `json:"id"`
+		ID          int64  `json:"id"`
 		Subject     string `json:"subject"`
 		Description string `json:"description"`
 	}
